refactor(condo): DOMA-9177 refactored helper utils
